fix: migrate golangci-lint configs to v2, fix lint errors across monorepo
- Rewrite .golangci.yaml for juno and callisto to golangci-lint v2 format
- Remove deprecated linters (deadcode, golint, maligned, scopelint, etc.)
- Use system golangci-lint in callisto Makefile instead of go run v1.55.2
- Remove unnecessary int64(tx.Height) casts after gas_wanted type change
- Fix gofmt/goimports issues in callisto
- Fix ESLint errors in UI chart component (|| → ??, unused vars)
- Fix ESLint errors in callisto adapter (no-base-to-string, no-unnecessary-condition)
- Restore README.md content

+++ apps/callisto/modules/auth/handle_msg.go
@@ -54,7 +54,7 @@ func (m *Module) HandleMsg(_ int, msg juno.Message, tx *juno.Transaction) error 
 		}
 	}
 
-	return m.RefreshAccounts(int64(tx.Height), moduleutils.FilterNonAccountAddresses(addresses))
+	return m.RefreshAccounts(tx.Height, moduleutils.FilterNonAccountAddresses(addresses))
 }
 
 func (m *Module) handleMsgCreateVestingAccount(msg *vestingtypes.MsgCreateVestingAccount, txTimestamp time.Time) error {
